+++ app/src/main/java/com/davidbugayov/financeanalyzer/utils/TestDataGenerator.kt
@@ -5,6 +5,7 @@ import com.davidbugayov.financeanalyzer.domain.model.Money
 import com.davidbugayov.financeanalyzer.domain.model.Transaction
 import com.davidbugayov.financeanalyzer.ui.theme.ExpenseColorInt
 import com.davidbugayov.financeanalyzer.ui.theme.IncomeColorInt
+import java.math.BigDecimal
 import java.util.Calendar
 import java.util.Date
 import kotlin.random.Random
@@ -82,7 +83,7 @@ object TestDataGenerator {
             transactions.add(
                 Transaction(
                     id = i.toString(),
-                    amount = Money(amount.toInt().toDouble()), // Округляем до целых
+                    amount = Money(BigDecimal.valueOf(amount)), // Используем BigDecimal без округления
                     category = category,
                     isExpense = isExpense,
                     date = date,
